+++ CHANGELOG.md
@@ -33,6 +33,7 @@ All notable changes to this project will be documented in this file.
   - Policy Definition Page
   - Create Policy Definition
   - Delete Policy Definition
+  - Dashboard Page
 - DCAT Catalog Service: Parse DCAT Catalog responses for use in the UI API Wrapper and the Broker Server.
 - JSON and JSON-LD Utilities: Centralized Vocab collection for dealing with EDC / DCAT JSON-LD